+++ src/onegov.org/onegov/org/path.py
@@ -15,7 +15,6 @@ from onegov.event import (
 from onegov.form import (
     FormDefinition,
     FormCollection,
-    FormSubmissionFile,
     CompleteFormSubmission,
     PendingFormSubmission
 )
@@ -144,12 +143,6 @@ def get_pending_form_submission(app, id):
 def get_complete_form_submission(app, id):
     return FormCollection(app.session()).submissions.by_id(
         id, state='complete', current_only=False)
-
-
-@OrgApp.path(model=FormSubmissionFile, path='/formular-datei/{id}',
-             converters=dict(id=UUID))
-def get_form_submission_file(app, id):
-    return FormCollection(app.session()).submissions.file_by_id(id)
 
 
 @OrgApp.path(model=Editor, path='/editor/{action}/{trait}/{page_id}')