+++ src/onegov.org/onegov/org/views/form_submission.py
@@ -1,6 +1,5 @@
 """ Renders and handles defined forms, turning them into submissions. """
 
-import base64
 import morepath
 
 from onegov.core.security import Public, Private
@@ -8,7 +7,6 @@ from onegov.ticket import TicketCollection
 from onegov.form import (
     FormCollection,
     FormDefinition,
-    FormSubmissionFile,
     PendingFormSubmission,
     CompleteFormSubmission
 )
@@ -191,12 +189,3 @@ def handle_complete_submission(self, request):
             request.success(_("Thank you for your submission!"))
 
             return morepath.redirect(request.link(ticket, 'status'))
-
-
-@OrgApp.view(model=FormSubmissionFile, permission=Private)
-def view_form_submission_file(self, request):
-    response = morepath.Response(base64.b64decode(self.filedata))
-    response.content_type = self.submission_data['mimetype']
-    response.content_encoding = 'gzip'
-
-    return response